refactor(workflows): extract asset loader to workflow-asset-loader seam (D5)
Separate pure loading/parsing concerns from run orchestration. Asset resolution
and document projection now live in dedicated workflow-asset-loader.ts, removing
database and file I/O ownership from runs.ts. This seam enables cleaner testing
and future optimization (caching, mocking) of asset resolution.
