+++ src/workflows/runtime/runs.ts
@@ -3,23 +3,15 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import { randomUUID } from "node:crypto";
-import fs from "node:fs";
 import { parseAssetRef } from "../../core/asset/asset-ref";
 import { loadConfig } from "../../core/config/config";
 import { NotFoundError, UsageError } from "../../core/errors";
 import { appendEvent } from "../../core/events";
-import { getDbPath } from "../../core/paths";
-import { closeDatabase, openExistingDatabase } from "../../indexer/db/db";
-import { resolveSourceEntries } from "../../indexer/search/search-source";
-import { resolveSourcesForOrigin } from "../../registry/origin-resolve";
-import { resolveAssetPath } from "../../sources/resolve";
 import type {
-  WorkflowParameter,
   WorkflowRunStatus,
   WorkflowRunStepState,
   WorkflowRunStepStatus,
   WorkflowRunSummary,
-  WorkflowStepDefinition,
 } from "../../sources/types";
 import {
   type WorkflowRunRow,
@@ -27,22 +19,11 @@ import {
   type WorkflowRunsRepository,
   withWorkflowRunsRepo,
 } from "../../storage/repositories/workflow-runs-repository";
-import { formatWorkflowErrors } from "../authoring/authoring";
 import { getCurrentWorkflowScopeKey } from "../authoring/scope-key";
-import { parseWorkflow } from "../parser";
-import type { WorkflowDocument } from "../schema";
 import { type SummaryJudge, validateStepSummary } from "../validate-summary";
 import { resolveAgentIdentity } from "./agent-identity";
 import { type CheckinDirective, evaluateCheckin } from "./checkin";
-
-type WorkflowAsset = {
-  ref: string;
-  path: string;
-  sourcePath: string;
-  title: string;
-  parameters?: WorkflowParameter[];
-  steps: WorkflowStepDefinition[];
-};
+import { loadWorkflowAsset, resolveWorkflowEntryId } from "./workflow-asset-loader";
 
 export interface WorkflowRunDetail {
   run: WorkflowRunSummary;
@@ -436,125 +417,6 @@ async function resolveRunSpecifier(
 
   const started = await startWorkflowRun(ref, params ?? {});
   return { run: readWorkflowRun(repo, started.run.id), autoStarted: true };
-}
-
-async function loadWorkflowAsset(ref: string): Promise<WorkflowAsset> {
-  const parsed = parseAssetRef(ref);
-  if (parsed.type !== "workflow") {
-    throw new UsageError(`Expected a workflow ref (workflow:<name>), got "${ref}".`);
-  }
-
-  const config = loadConfig();
-  const allSources = resolveSourceEntries(undefined, config);
-  const searchSources = resolveSourcesForOrigin(parsed.origin, allSources);
-  let assetPath: string | undefined;
-  let sourcePath: string | undefined;
-
-  for (const source of searchSources) {
-    try {
-      assetPath = await resolveAssetPath(source.path, "workflow", parsed.name);
-      sourcePath = source.path;
-      break;
-    } catch {
-      /* continue */
-    }
-  }
-
-  if (!assetPath) {
-    throw new NotFoundError(`Workflow not found for ref: workflow:${parsed.name}`);
-  }
-
-  const resolvedSourcePath = sourcePath ?? config.stashDir ?? assetPath;
-  const fullRef = `${parsed.origin ? `${parsed.origin}//` : ""}workflow:${parsed.name}`;
-
-  const cached = readWorkflowDocumentFromIndex(resolvedSourcePath, fullRef);
-  const document = cached ?? loadWorkflowDocumentFromDisk(assetPath);
-  return projectAsset(document, fullRef, assetPath, resolvedSourcePath);
-}
-
-function loadWorkflowDocumentFromDisk(assetPath: string): WorkflowDocument {
-  const content = fs.readFileSync(assetPath, "utf8");
-  const result = parseWorkflow(content, { path: assetPath });
-  if (!result.ok) {
-    throw new UsageError(formatWorkflowErrors(assetPath, result.errors));
-  }
-  return result.document;
-}
-
-function readWorkflowDocumentFromIndex(sourcePath: string, ref: string): WorkflowDocument | null {
-  const dbPath = getDbPath();
-  if (!fs.existsSync(dbPath)) return null;
-
-  const db = openExistingDatabase(dbPath);
-  try {
-    const parsed = parseAssetRef(ref);
-    const entryKey = `${sourcePath}:${parsed.type}:${parsed.name}`;
-    const row = db
-      .prepare(
-        `SELECT wd.document_json AS document_json
-           FROM workflow_documents wd
-           JOIN entries e ON e.id = wd.entry_id
-          WHERE e.entry_type = 'workflow' AND e.entry_key = ?
-          LIMIT 1`,
-      )
-      .get(entryKey) as { document_json: string } | undefined;
-    if (!row) return null;
-    try {
-      return JSON.parse(row.document_json) as WorkflowDocument;
-    } catch {
-      return null;
-    }
-  } finally {
-    closeDatabase(db);
-  }
-}
-
-function projectAsset(doc: WorkflowDocument, ref: string, assetPath: string, sourcePath: string): WorkflowAsset {
-  return {
-    ref,
-    path: assetPath,
-    sourcePath,
-    title: doc.title,
-    ...(doc.parameters
-      ? {
-          parameters: doc.parameters.map((p) => ({
-            name: p.name,
-            ...(p.description ? { description: p.description } : {}),
-          })),
-        }
-      : {}),
-    steps: doc.steps.map((s) => ({
-      id: s.id,
-      title: s.title,
-      instructions: s.instructions.text,
-      ...(s.completionCriteria ? { completionCriteria: s.completionCriteria.map((c) => c.text) } : {}),
-      sequenceIndex: s.sequenceIndex,
-    })),
-  };
-}
-
-function resolveWorkflowEntryId(sourcePath: string, ref: string): number | null {
-  const dbPath = getDbPath();
-  if (!fs.existsSync(dbPath)) return null;
-
-  const db = openExistingDatabase(dbPath);
-  try {
-    const parsed = parseAssetRef(ref);
-    const entryKey = `${sourcePath}:${parsed.type}:${parsed.name}`;
-
-    const row = db
-      .prepare(
-        `SELECT id
-         FROM entries
-         WHERE entry_type = 'workflow'
-            AND entry_key = ?
-          LIMIT 1`,
-      )
-      .get(entryKey) as { id: number } | undefined;
-    return row?.id ?? null;
-  } finally {
-    closeDatabase(db);
-  }
 }
 
 function readWorkflowRun(repo: WorkflowRunsRepository, runId: string): WorkflowRunRow {
